Save money by letting a slave run until the paid hour runs out
If the idle time is set to a negative number, the slave won't be
destroyed until it's in the idle state and it's about to get charged
for the next hour by Digital Ocean. Digital Ocean charges for the
first hour whenever a droplet is created, so creating 10 droplets that
run for 5 minutes each before getting destroyed will result in a bill
for 10 hours. However, if you had created a single droplet running for
50 minutes, you'd be billed just for 1 hour. So it makes sense to have
an option wich, instead of keeping an idle slave around for some idle
time, will keep it around until right before it would get charged for
the next hour, reducing the number of droplet recreations, thus saving
money.

Note that Digital Ocean charges for the second hour at 1 hour and 30
minutes as they do rounding (1h30m is 1.5 hours, which rounds to 2).
However, the third, fourth and so on hours are 1 hour long (2h30m,
3h30m, etc.), so it's only the first hour which is 1h30m long.

Fun thought: this means that if you keep rectreating your droplet right
before it hits 1h30m, all your hours will be the first hours, and if
you keep it up for the duration of a month, your monthly droplet payment
will be 2/3 (66.6%) of the one on the Pricing page (e.g. you will be
paying $26.6 for the $40 droplet).